ADD Map of tables

+++ app/Http/Controllers/Store/BoardController.php
@@ -14,6 +14,7 @@ use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Response;
 use App\Models\Board;
+use App\Models\BoardPosition;
 use App\Models\Store;
 use Illuminate\Database\Eloquent\Builder;
 
@@ -21,7 +22,8 @@ class BoardController extends Controller
 {
     public function index()
     {
-        $boards = Board::orderBy('id')->get();
+        $store = $this->storeInSesion();
+        $boards = Board::where('fk_id_store', $store->id)->with('position')->get();
         return view('store.board.index', ['boards' => $boards]);
     }
 
@@ -35,17 +37,14 @@ class BoardController extends Controller
     {
         $board = Board::find($boardId);
         return view('store.board.profile', ['board' => $board]);
-        
+
     }
-    
+
 
     public function upsertPost(UpsertBoardRequest $request ,$boardId = 0)
     {
         try{
-            $user = Auth::User();
-        $store = Store::whereHas('employees', function (Builder $query) use ($user) {
-            $query->where('fk_id_user', $user->id);
-        })->first();
+            $store = $this->storeInSesion();
             DB::beginTransaction();
             $board = Board::findOrNew($boardId);
             $board->fill($request->all());
@@ -53,7 +52,7 @@ class BoardController extends Controller
             $board->saveOrfail();
             DB::commit();
             return redirect()->route('store_board_index');
-     } catch(\Exception $e) {
+        } catch(\Exception $e) {
             DB::rollback();
             return back()
             ->withErrors([ 'generic_error' => ['Algo salio mal']])
@@ -64,7 +63,38 @@ class BoardController extends Controller
     public function destroy($boardId)
     {
         $board = Board::find($boardId);
+        $store = $this->storeInSesion();
+        if($board->fk_id_store != $store->id){
+            return back()->withErrors([ 'generic_error' => ['Datos no encontrados']]);
+        }
         $board->delete();
         return redirect()->route('store_board_delete');
+    }
+
+    public function map()
+    {
+        $store = $this->storeInSesion();
+        $boards = Board::where('fk_id_store', $store->id)->with('position')->get();
+        return view('store.board.map', ['boards' => $boards]);
+    }
+
+    public function update(Request $request)
+    {
+        $success = false;
+        $store = $this->storeInSesion();
+        $positions = $request->get('positions');
+        BoardPosition::whereHas('board', function (Builder $query) use ($store) {
+            $query->where('fk_id_store', $store->id);
+        })->delete();
+
+        foreach($positions as $position){
+            $boardPosition = new BoardPosition();
+            $boardPosition->x = $position["x"];
+            $boardPosition->y = $position["y"];
+            $boardPosition->fk_id_board = $position["id"];
+            $boardPosition->saveOrFail();
+        }
+        $success = true;
+        return response()->json(['success' => $success]);
     }
 }
